Support object types from other parcels without prefix
Resolve types after all classes have been parsed.

This means that caching of C representations has to be postponed. For
now, the C representations are cached after the types have been
resolved. This is error-prone and requires some hacks for methods. It
might be better to move to a scheme where the C strings are not cached
until they're used for the first time.

TODO: Add Perl bindings and fix Perl tests