Rework Clownfish test harness
This commit corrects some bad design decisions I made when I
reimplemented the Clownfish test harness.

The new TestBatch class that the test modules inherit from contains only
a single abstract method "Run". All the other functionality is moved to
a new class TestBatchRunner. An instance of TestBatchRunner is passed to
the "Run" method of TestBatch. This means that TestBatch doesn't have to
care about TestFormatters.

The old TestRunner is renamed to TestSuiteRunner. There's a new class
TestSuite that manages a collection of TestBatches.